feat(tests): add tests/ folder to better structure
